Change temp to int16_t and fix operator precedence

+++ src/include/protocol_sensors.h
@@ -60,7 +60,7 @@ struct SensorIMU {
 } __attribute__((packed));
 
 struct SensorTemp {
-    int temp;
+    int16_t temp;
 }__attribute__((packed));
 
 struct SensorADCS {
